Changed when old backups are deleted, and made it so if fullbackup fails it can do a full on rerun

+++ mariabackup.bash
@@ -6,8 +6,6 @@
 #create user 'backup'@'localhost' identified by 'password';
 #grant reload,process,lock tables,binlog monitor,connection admin,slave monitor on *.* to 'backup'@'localhost';
 
-#------customize main settings-------
-
 # Define the backup directory
 backup_dir=/media/backups/
 
@@ -15,9 +13,9 @@ backup_dir=/media/backups/
 user=backup
 password=password
 
-#email list, spaces in-between, no commas
-emails="email@emaildomain.com"
-fromemail="mariabackup@emaildomain.com"
+#emaillist, spaces in-between, no commas
+emails="harry.pask@hotmail.com"
+fromemail="mariabackupalerts@6dglinux.com"
 
 #number of days to keep backups
 #0= just today's backup | 1= today and yesterday | 2=today,yesterday,day before etc
@@ -26,15 +24,15 @@ backupdays=0
 #----------define backup options------------
 #incremental options
 declare -a backup_options_inc=(
-	"--backup"
-	"--user=$user"
-	"--password=$password"
-	"--extra-lsndir=$extra_lsndir"
-	"--incremental-basedir=$extra_lsndir"
-	"--stream=xbstream"
-	"--slave-info"
-	"--parallel=1"
-	)
+		"--backup"
+		"--user=$user"
+		"--password=$password"
+		"--extra-lsndir=$extra_lsndir"
+		"--incremental-basedir=$extra_lsndir"
+		"--stream=xbstream"
+		"--slave-info"
+		"--parallel=1"
+		)
 
 #full backup options
 declare -a backup_options_full=(
@@ -44,8 +42,8 @@ declare -a backup_options_full=(
         "--target-dir=$fullbackuplocation"
         "--extra-lsndir=$extra_lsndir"
         "--stream=xbstream"
-	"--slave-info"
-	"--parallel=1"
+		"--slave-info"
+		"--parallel=1"
         )
 
 #------------variables------------
@@ -67,12 +65,6 @@ incremental_folder=$current_date_folder/incr/$current_datetime
 fullbackuplocation=$current_date_folder/fullbackup
 mkdir -p $current_date_folder
 
-#wait so old backup has time to "age"
-#sleep 10
-
-# Delete folders in backup directory older than 24 hours
-find $backup_dir -maxdepth 1 -name '20*' -mtime +$backupdays | xargs rm -rf
-
 #------backup process-------
 cd $current_date_folder
 # Check if full backup file exists
@@ -84,7 +76,7 @@ else
 	# Perform full backup and create create $full_backup_file
 	mkdir -p $fullbackuplocation
 	mariabackup "${backup_options_full[@]}" 2>> $current_date_folder/backup.log | gzip > $fullbackuplocation/full_backup.gz
-	touch $full_backup_file
+	#touch $full_backup_file
 	#mv $current_date_folder/full_backup.gz $fullbackuplocation
 fi
 
@@ -95,15 +87,23 @@ checkstatus=$(tail -n 2 /$current_date_folder/backup.log| grep -c "completed OK"
 
 #if completed OK! is at the end of the backup log file then add status to backup_status.log
 #if not then send backup.log to email address and delete failed incremental backup folder so prepare script doesn't break
+#If full backup is successful then make $full_backup_file, if fullbackup fails then file is not created so on next run a fullbackup is tried again
 
 if [[ $checkstatus -eq 1 ]]; then
     echo "$(date +'%Y-%m-%d %H:%M:%S') mariabackup completed okay" >> $current_date_folder/backup_status.log
-
+	touch $full_backup_file
+	find $backup_dir -maxdepth 1 -name '20*' -mtime +$backupdays | xargs rm -rf
 else
     log_content=$(tail -n 200 "$current_date_folder/backup.log") 
     echo "$log_content" | mailx -r $fromemail -s "MariaBackup task for $HOSTNAME failed" $emails
-    rm -rf $incremental_folder
-	echo "$(date +'%Y-%m-%d %H:%M:%S') mariabackup failed - file $incremental_folder deleted" >> $current_date_folder/backup_status.log
-fi
+		#checks if full backup completed, if it has then failed incremental backup is removed
+		if [[ -f $full_backup_file ]]; then
+			rm -rf $incremental_folder
+			echo "$(date +'%Y-%m-%d %H:%M:%S') mariabackup failed - file $incremental_folder deleted" >> $current_date_folder/backup_status.log
+		else 
+			echo "$(date +'%Y-%m-%d %H:%M:%S') Full backup failed, please reslove issue and rerun backup" >> $current_date_folder/backup_status.log
+		fi
+fi	
+
